GEODE-5025: Move templates internal.
