+++ cppcache/src/SerializationRegistry.hpp
@@ -29,6 +29,7 @@
 
 #include <geode/internal/geode_globals.hpp>
 #include <geode/internal/DataSerializableInternal.hpp>
+#include <geode/internal/DataSerializablePrimitive.hpp>
 #include <geode/Serializable.hpp>
 #include <geode/PdxSerializer.hpp>
 #include <geode/DataOutput.hpp>
@@ -61,6 +62,7 @@ namespace geode {
 namespace client {
 
 using internal::DataSerializableInternal;
+using internal::DataSerializablePrimitive;
 
 typedef ACE_Hash_Map_Manager<int64_t, TypeFactoryMethod, ACE_Null_Mutex>
     IdToFactoryMap;
